Move SparkRunnerClassLoader into classloader package
