+++ cdap-spark-core/src/test/java/co/cask/cdap/app/runtime/spark/classloader/SparkRunnerClassLoaderTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 Cask Data, Inc.
+ * Copyright © 2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,8 +14,9 @@
  * the License.
  */
 
-package co.cask.cdap.app.runtime.spark;
+package co.cask.cdap.app.runtime.spark.classloader;
 
+import co.cask.cdap.app.runtime.spark.classloader.SparkRunnerClassLoader;
 import co.cask.cdap.common.lang.ClassLoaders;
 import com.google.common.io.Closeables;
 import org.junit.Test;
